move runFirstColumnWrites to Cypher class
